`GitHubApiLoggingInterceptor` 경로 변경
- `infrastructure.resttemplate.interceptor` -> `infrastructure.resttemplate.github.interceptor`
